eventListCard style final
