+++ frontend/src/components/EventList/EventListCard.styles.js
@@ -1,0 +1,38 @@
+const styles = {
+    main_grid: {
+        "margin-top": 15,   
+    },
+    card: {
+        minWidth: 275,
+        height: 185,  
+    },
+    bullet: {
+        display: 'inline-block',
+        margin: '0 2px',
+        transform: 'scale(0.8)',
+    },
+    title: {
+        fontFamily: "Raleway",
+        fontSize: "30px",
+        fontWeight: "700",
+    },
+    pos: {
+        marginBottom: 12,
+    },
+    icon: {
+        fontSize: 28,
+        color: "#B22222",
+        verticalAlign: "text-bottom",
+    },
+    coord: {
+        fontFamily: "Raleway",
+        fontSize: "16px",
+        fontWeight: "700",
+    },
+    description: {
+        fontFamily: "Raleway",
+    }
+    
+}
+
+export default styles;
